Fix error: call to 'abs' is ambiguous
An unsigned int minus a signed int yields an unsigned int.
Calling abs on an unsigned int is ambiguous.
Cast the unsigned int to a signed int before calling abs.

+++ src/Algorithm/GapFillProcess.cpp
@@ -233,7 +233,7 @@ GapFillReturnCode GapFillProcess::selectGapSequence(int estimatedSize, const Str
 
     for(size_t i = 0; i < sequences.size(); ++i)
     {
-        int diff = abs(sequences[i].size() - estimatedSize);
+        int diff = abs((int)sequences[i].size() - estimatedSize);
         //printf("ES: %d S: %zu D: %d\n", estimatedSize, sequences[i].size(), diff);
 
         if(diff < selectedSizeDiff)
